# Make Genus Faster While Increasing Clock Frequency