Lowered density of MEM, changed constraints very slightly, and decreased clock period to 1.1ns. PT Timing signoff reports 0.008 WNS, and total area of MEM tile is 16.8K. Lowering density allows entire flow up until cadence-innovus-signoff to complete in 4h10min on arm. Keeping density at 0.74 with 1.1ns clock period is an easy to way to let postroute take a _ very _ long time.

Got rid of modes in PE genus synthesis node, which was slowing down synthesis times a lot. At 4ns, Genus went down from 30min to 6min. At 1.1ns, Genus takes 19min, and entire flow up until cadence-innovus-signoff takes 4hr49min on arm. No timing violations found, and total area (with pond) is 5907. With pond at 4ns, area was 5460.

These times were tested on arm. Let me know if I should run on VDE to make sure that the runtimes are also reasonable on the cloud.