@alexcarsello 2020-10-25
Nice! Can you provide the path to the build directories for each tile? I want to take a look and confirm the tile array will still fit on the chip.