@norabarlow 2020-10-25
`/sim/kzf/garnet/garnet/mflowgen/Tile_PE/build_timing_1.1ns_wpond`
`/sim/kzf/garnet/garnet/mflowgen/Tile_MemCore/build_timing_1.1ns`