@norabarlow 2020-10-25
LVS is good for both. A few DRC warnings/errors for both as well.