@alexcarsello 2020-10-25
It's close, but it looks like it will fit on chip.